requirements.txt
boto3==1.26.23
psycopg2-binary==2.9.5
pandas==1.5.2
numpy==1.23.5
pytz==2022.6
sql-query==0.4.1

